Fixed issue that was blocking log in on the server.
It seems like the index was being wrongly get from the packet CB_START_GAME.

+++ src/BarracksServer/Network/PacketHandler.cs
@@ -414,7 +414,7 @@ namespace Melia.Barracks.Network
 		public void CB_START_GAME(IBarracksConnection conn, Packet packet)
 		{
 			var channelId = packet.GetShort();
-			var characterIndex = packet.GetShort();
+			var characterIndex = packet.GetByte();
 
 			// Get character
 			var character = conn.Account.GetCharacterByIndex(characterIndex);
